fix: refactor and prevent daily to override the manual import
